added conftest.py for pytest configuration
